requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Django==3.1.2
django-bootstrap-form==3.4
django-celery-beat==2.1.0
django-celery-results==1.2.1
django-filter==2.4.0
django-stdimage==5.1.1
django-timezone-field==4.0
djangorestframework==3.12.1
kombu==4.6.10
Markdown==3.3.3
Pillow==8.0.1
progressbar2==3.53.1
prompt-toolkit==3.0.8
python-crontab==2.5.1
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2020.1
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
vine==1.3.0
wcwidth==0.2.5
